pyproject.toml
[project]
name = "ffxiv-piper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "onnxruntime-gpu>=1.22.0",
    "piper-tts>=1.3.0",
    "sounddevice>=0.5.2",
]
